+++ src/appsettings.cpp
@@ -32,6 +32,9 @@
 #include "singleapplication.h"
 
 QTranslator AppSettings::m_appTranslator;
+#if defined(Q_OS_WIN)
+QTranslator AppSettings::m_qtTranslator;
+#endif
 
 AppSettings::AppSettings(QObject *parent) :
     QSettings(parent)
@@ -42,6 +45,9 @@ void AppSettings::setupLocale()
 {
     loadLocale(locale());
     SingleApplication::installTranslator(&m_appTranslator);
+#if defined(Q_OS_WIN)
+    SingleApplication::installTranslator(&m_qtTranslator);
+#endif
 }
 
 QLocale::Language AppSettings::locale()
@@ -65,6 +71,9 @@ void AppSettings::loadLocale(QLocale::Language lang)
         QLocale::setDefault(QLocale(lang));
 
     m_appTranslator.load(QLocale(), "crow", "_", ":/translations");
+#if defined(Q_OS_WIN)
+    m_qtTranslator.load(QLocale(), "qt", "_", QLibraryInfo::location(QLibraryInfo::TranslationsPath));
+#endif
 }
 
 AppSettings::WindowMode AppSettings::windowMode()