+++ src/appsettings.h
@@ -243,6 +243,9 @@ public:
 
 private:
     static QTranslator m_appTranslator;
+#if defined(Q_OS_WIN)
+    static QTranslator m_qtTranslator; // Qt library translations
+#endif
 };
 
 #endif // APPSETTINGS_H