+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - Fix selected language on search
+- Windows: Load Qt library translations
 
 ## [2.0.0] - 2018-11-17
 ### Added